+++ R/Rcopy.R
@@ -1,9 +1,9 @@
-#' Copy and open R and Rmd files with QMS approved headers.
+#' Copy and open R, Rmd, and qmd files with QMS approved headers.
 #'
 #' This function is somewhat similar to the Cognigen shell command \code{Rcopy},
 #' but it is not called.
 #'
-#' @param from path of original R program or Rmd file. Defaults to the path of
+#' @param from path of original R, Rmd, or qmd file. Defaults to the path of
 #'   the source editor context.
 #'   
 #' @param to path or directory of new file. If \code{to} is a directory,
@@ -85,14 +85,14 @@ Rcopy <- function(from = NULL,
   file_ext_to <- tolower(tools::file_ext(to))
 
   assertthat::assert_that(
-    file_ext_from %in% c("r", "rmd"),
-    msg = "`from` must be an R or Rmd file"
+    file_ext_from %in% c("r", "rmd", "qmd"),
+    msg = "`from` must be an R, Rmd, or qmd file"
   )
 
   assertthat::assert_that(
-    file_ext_to %in% c("r", "rmd"),
+    file_ext_to %in% c("r", "rmd", "qmd"),
     !file.exists(to),
-    msg = "`to` must be a non-existing R or Rmd file"
+    msg = "`to` must be a non-existing R, Rmd, or qmd file"
   )
 
   assertthat::assert_that(
@@ -104,7 +104,7 @@ Rcopy <- function(from = NULL,
 
   assertthat::assert_that(
     file.exists(from),
-    msg = "`from` must be an existing R or Rmd file"
+    msg = "`from` must be an existing R, Rmd, or qmd file"
   )
   
   old_header <- get_header(from)
@@ -123,7 +123,7 @@ Rcopy <- function(from = NULL,
   }
 
   # how files are copied depends on their type
-  if(file_ext_to == "rmd") {
+  if(file_ext_to %in% c("rmd", "qmd")) {
 
     if(isFALSE(old_header)) {